move last dvv function from noise_module
